pyproject.toml
[project]
name = "vibe-opencode-pocketsmith-beancount"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beancount>=3.1.0",
    "colorama>=0.4.6",
    "pocketsmith-api>=2.1.1",
    "python-dotenv>=1.1.1",
    "pytz>=2024.2",
    "PyYAML>=6.0",
    "regex>=2023.0.0",
    "requests>=2.32.4",
    "typer>=0.9.0",
]

[dependency-groups]
dev = [
    "hypothesis[datetime]>=6.100.0",
    "mypy>=1.17.1",
    "pre-commit>=4.2.0",
    "psutil>=5.9.0",
    "pytest>=8.4.1",
    "pytest-cov>=4.0.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.12.7",
    "types-PyYAML>=6.0.0",
    "types-pytz>=2024.2.0.20241003",
    "types-requests>=2.32.4.20250611",
]
